Fix usage of Schema Voyager from external projects
The plan is to make Schema Voyager work as a a dependency. That is, you
can include it in your projects deps, then execute functions in it, and
have the standalone HTML output in your own repo.

This is basically working now, however...

For it to really work, it needs a large (~800k, or ~200k gzipped)
template file, into which the schema DB can be injected. I'm keeping
this out of the repo for now. I may commit it eventually, but will look
at other means of distributing it first.