Fix cljdoc by splitting up project
There are now 3 pieces to the project:
1. The core, (src/core) which includes tools for ingesting data from sources,
   converting that into a DataScript DB, and building a standalone HTML
   page from the template.
2. The web, (src/web) which includes the CLJS for rendering the data in
   the DataScript DB.
3. The build (src/build) which includes the scripts for creating the
   template.

Each piece has its own dependencies. This is important because CLJDOC is
very touchy about dependencies, and will only work if all the
dependencies are in the pom. Splitting everything up like this keeps the
dependency list small in the POM, makes the JAR small, and lets CLJDOC
analyze only src/core, which is where the public api
(schema-voyager/cli) lives.

I have further plans for restructuring the namespaces, but that can come
in a later commit.