Check CDN integrity
Switches graphviz CDNs to the one currently recommended by hpcc-js.

Also, as per best-practices when using CDNs, adds an integrity attribute
to the script tag, to ensure the integrity of the main hpcc-js file. Not
sure if this also ensures the integrity of the .wasm file that gets
downloaded by the main script file.

https://developer.mozilla.org/en-US/docs/Web/Security/Subresource_Integrity